+++ src/components/Travelogue/FeedHeader.tsx
@@ -1,7 +1,8 @@
-import { Avatar, Stack, Typography } from '@mui/material';
+import { Stack, Typography } from '@mui/material';
 import Image from 'next/image';
-import { avatar, location } from 'public/images';
+import { location } from 'public/images';
 
+import UserProfile from '@/components/common/UserProfile';
 import { TravelogueFeedType } from '@/types/travelogue';
 
 type FeedHeaderProps = TravelogueFeedType['member'] & {
@@ -19,18 +20,7 @@ const FeedHeader = ({
       justifyContent='space-between'
       my={0.8}
       sx={{ height: '30px' }}>
-      <Stack direction='row' alignItems='center' spacing={0.8}>
-        <Avatar
-          src={profileImageUrl !== 'default' ? profileImageUrl : undefined}
-          sx={{ bgcolor: 'gray020.main', width: 28, height: 28, textAlign: 'center' }}>
-          {profileImageUrl === 'default' && <Image src={avatar} width={28} alt='' />}
-        </Avatar>
-        <Typography
-          component='span'
-          sx={{ color: 'dark.main', fontSize: '16px', ...fontStyle }}>
-          {nickname}
-        </Typography>
-      </Stack>
+      <UserProfile profileImageUrl={profileImageUrl} nickname={nickname} />
       <Stack
         direction='row'
         alignItems='center'
@@ -47,7 +37,7 @@ const FeedHeader = ({
         <Image src={location} width={12} alt='' />
         <Typography
           component='span'
-          sx={{ ...fontStyle, color: 'blue050.main', fontSize: '14px' }}>
+          sx={{ fontWeight: '500', pb: '3px', color: 'blue050.main', fontSize: '14px' }}>
           {country}
         </Typography>
       </Stack>
@@ -56,8 +46,3 @@ const FeedHeader = ({
 };
 
 export default FeedHeader;
-
-const fontStyle = {
-  fontWeight: '500',
-  pb: '3px',
-};
